% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindPartitions}
\alias{FindPartitions}
\title{Finds partitions akin to Monocle3 FindPartition Function}
\usage{
FindPartitions(
  obj = seurat,
  method = "louvain",
  k = 20,
  reduction = "pca",
  dims = NULL,
  weight = F,
  num_iter = 1,
  resolution_parameter = NULL,
  random_seed = 2020,
  verbose = T,
  partition_q_value = 0.05
)
}
\arguments{
\item{obj}{Input seurat object.}

\item{method}{Louvain or Leiden}

\item{k}{nearest number k}

\item{reduction}{reduction to find partitions of}

\item{dims}{number of dimensions}

\item{num_iter}{Number of iterations (leiden only); Default is 1.}

\item{resolution_parameter}{leiden only}

\item{random_seed}{seed default 2020}

\item{verbose}{Default True}

\item{partition_q_value}{Louvain only - sets resolution}
}
\description{
Stated differently, low cluster resolution
}
